requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.2.2
google-cloud-vision==3.4.4
numpy>=1.20.0
pydantic==1.10.7
faiss-cpu==1.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
